+++ lib/main_db/core/daos/crud/crud_types.dart
@@ -1,0 +1,3 @@
+import 'package:hoplixi/main_db/core/main_store.dart';
+
+typedef VaultItemWith<T extends Object> = (VaultItemsData, T);